Migrated the "responsible AI" integration tests to GitHub Models and simplified them
